update-2h30

+++ server/app/Http/Controllers/API/BlogController.php
@@ -17,18 +17,38 @@ class BlogController extends Controller
         return BlogResource::collection($blog);
     }
     public function store(CreateBlogRequest $request){
-        
-        $data = $request->all();
-        $blog = Blog::create($data);
-        return MessageStatusAPI::store($blog);
+        $user_id = auth()->user()->id;
+        $validated = $request->validated();
+        $blog = new Blog([
+            'title' => $validated['title'],
+            'content' => $validated['content'],
+            'user_id' => $user_id,
+        ]);
+        $blog->save();
+        return MessageStatusAPI::store();
+       
     }
     public function update(CreateBlogRequest $request, $id){
         $blog = Blog::find($id);
+        $user_id = auth()->user()->id;
+        if(!$user_id){
+            return MessageStatusAPI::displayInvalidInput($blog);
+        }
+        if($user_id !== $blog->user_id){
+            return MessageStatusAPI::notFound();
+        }
         $blog->update($request->all());
         return MessageStatusAPI::update();
     }
     public function destroy($id){
-        $blog = Blog::findOrFail($id);
+        $blog = Blog::find($id);
+        $user_id = auth()->user()->id;
+        if(!$user_id){
+            return MessageStatusAPI::displayInvalidInput($blog);
+        }
+        if($user_id !== $blog->user_id){
+            return MessageStatusAPI::notFound();
+        }
         $blog->delete();
         return MessageStatusAPI::destroy();
     }
